Fix for NVMe temp sensor lost during power cycle (#16)
Currently NVMe MI app takes copy of all the drive objects to do health status polling. This extends the life of thos objects until polling of all the drives are completed. This creates error while trying to delete and create a drive object while polling is ongoing. Use weak_ptr instead of shared_ptr to extend the lifetime of drive objects only if it os not deleted.

And according to MCTP over I3C specification, maximum time-out waiting for a response should be 300ms. So changed hsPollTimeout to 300ms

Signed-off-by: ndandex <nikhilax.dande@intel.com>